+++ NEWS.md
@@ -1,3 +1,5 @@
+# cpp11 (development version)
+
 # cpp11 0.4.1
 
 * Fix crash related to unwind protect optimization (#244)