+++ NEWS.md
@@ -1,5 +1,7 @@
 # cpp11 (development version)
 
+* Type files `<basename>_type.h` is no honored by `source_cpp()` (#216). 
+
 # cpp11 0.4.1
 
 * Fix crash related to unwind protect optimization (#244)