+++ NEWS.md
@@ -1,6 +1,7 @@
 # cpp11 (development version)
 
 * Type files `<basename>_type.h` is no honored by `source_cpp()` (#216). 
+* The directory of the sourced file is included into the include path during `source_cpp()` compilation. 
 
 # cpp11 0.4.1