+++ NEWS.md
@@ -1,5 +1,7 @@
 # cpp11 (development version)
 
+* Romain François is now the maintainer.
+
 # cpp11 0.4.1
 
 * Fix crash related to unwind protect optimization (#244)