+++ NEWS.md
@@ -1,4 +1,4 @@
-# cpp11 (development version)
+# cpp11 0.4.2
 
 * Romain François is now the maintainer.